+++ src/backend/core/urls.py
@@ -12,6 +12,7 @@ from core.external_api import viewsets as external_api_viewsets
 # - Main endpoints
 router = DefaultRouter()
 router.register("items", viewsets.ItemViewSet, basename="items")
+router.register("mounts", viewsets.MountViewSet, basename="mounts")
 router.register("share-links", viewsets.ShareLinkViewSet, basename="share_links")
 router.register("users", viewsets.UserViewSet, basename="users")